+++ rubato/classes/components/rigidbody.py
@@ -9,7 +9,7 @@ from . import Component
 from ... import Vector, Defaults, Time
 
 if TYPE_CHECKING:
-    from . import ColInfo
+    from . import Manifold
 
 
 class RigidBody(Component):
@@ -42,7 +42,7 @@ class RigidBody(Component):
         """
         params = Defaults.rigidbody_defaults | options
 
-        super().__init__(params)
+        super().__init__()
 
         self.static: bool = params["static"]
 
@@ -53,6 +53,7 @@ class RigidBody(Component):
         self.pos_correction: float = params["pos_correction"]
 
         self.velocity: Vector = params["velocity"]
+        self.ang_vel: float = params["ang_vel"]
 
         self.singular = True
 
@@ -61,15 +62,42 @@ class RigidBody(Component):
         else:
             self.inv_mass: float = 1 / params["mass"]
 
+        if params["moment"] == 0 or self.static:
+            self.inv_moment = 0
+        else:
+            self.inv_moment: float = 1 / params["moment"]
+
         self.bounciness: float = params["bounciness"]
 
     @property
     def mass(self) -> float:
-        """The mass of the Rigidbody (readonly)"""
+        """The mass of the Rigidbody."""
         if self.inv_mass == 0:
             return 0
         else:
             return 1 / self.inv_mass
+
+    @mass.setter
+    def mass(self, new: float):
+        if new == 0:
+            self.inv_mass = 0
+        else:
+            self.inv_mass: float = 1 / new
+
+    @property
+    def moment(self) -> float:
+        """The moment of inertia of the Rigidbody."""
+        if self.inv_moment == 0:
+            return 0
+        else:
+            return 1 / self.inv_moment
+
+    @moment.setter
+    def moment(self, new: float):
+        if new == 0:
+            self.inv_moment = 0
+        else:
+            self.inv_moment: float = 1 / new
 
     def physics(self):
         """Applies general kinematic laws to the rigidbody."""
@@ -78,6 +106,7 @@ class RigidBody(Component):
         self.velocity.clamp(-self.max_speed, self.max_speed)
 
         self.gameobj.pos += self.velocity * Time.milli_to_sec(Time.fixed_delta)
+        self.gameobj.rotation += self.ang_vel * Time.milli_to_sec(Time.fixed_delta)
 
     def add_force(self, force: Vector):
         """
@@ -126,7 +155,7 @@ class RigidBody(Component):
         )
 
     @staticmethod
-    def handle_collision(col: ColInfo):
+    def handle_collision(col: Manifold):
         """
         Resolve the collision between two rigidbodies.
         Utilizes a simplistic impulse resolution method.
@@ -138,87 +167,111 @@ class RigidBody(Component):
         rb_a: RigidBody = col.shape_b.gameobj.get(RigidBody)
         rb_b: RigidBody = col.shape_a.gameobj.get(RigidBody)
 
-        rb_a_none = rb_a is None
-        rb_b_none = rb_b is None
+        a_none = rb_a is None
+        b_none = rb_b is None
 
-        if rb_a_none and rb_b_none:
+        if a_none and b_none:
             return
 
         # Find inverse masses
-        inv_mass_a: float = 0 if rb_a_none else rb_a.inv_mass
-        inv_mass_b: float = 0 if rb_b_none else rb_b.inv_mass
+        inv_mass_a: float = 0 if a_none else rb_a.inv_mass
+        inv_mass_b: float = 0 if b_none else rb_b.inv_mass
 
         # Handle infinite mass cases
         if inv_mass_a == inv_mass_b == 0:
-            if rb_a_none:
+            if a_none:
                 inv_mass_b = 1
-            elif rb_b_none:
+            elif b_none:
                 inv_mass_a = 1
             else:
                 inv_mass_a, inv_mass_b = 1, 1
 
         inv_sys_mass = 1 / (inv_mass_a + inv_mass_b)
 
-        # Find collision separation normal
-        collision_norm = col.sep.unit()
+        des_a_v = Vector()
+        des_b_v = Vector()
+        des_a_a = 0
+        des_b_a = 0
+
+        for contact in col.contacts:
+            # Impulse Resolution
+            ra = None if a_none else contact - rb_a.gameobj.pos
+            rb = None if b_none else contact - rb_b.gameobj.pos
+
+            # Relative velocity
+            rv = (0 if b_none else rb_b.velocity + rb.unit().perpendicular(rb_b.ang_vel)
+                 ) - (0 if a_none else rb_a.velocity + ra.unit().perpendicular(rb_a.ang_vel))
+
+            if (vel_along_norm := rv.dot(col.normal)) > 0:
+                continue
+
+            # Calculate restitution
+            e = max(0 if a_none else rb_a.bounciness, 0 if b_none else rb_b.bounciness)
+
+            # Calculate inverse angular components
+            i_a = 0 if a_none else ra.cross(col.normal)**2 * rb_a.inv_moment
+            i_b = 0 if a_none else rb.cross(col.normal)**2 * rb_b.inv_moment
+
+            # Calculate impulse scalar
+            j = -(1 + e) * vel_along_norm / (inv_mass_a + inv_mass_b + i_a + i_b)
+            j /= len(col.contacts)
+
+            # Apply the impulse
+            impulse = j * col.normal
+
+            if not (a_none or rb_a.static):
+                des_a_v += inv_mass_a * impulse
+                des_a_a += ra.cross(impulse) * rb_a.inv_moment
+
+            if not (b_none or rb_b.static):
+                des_b_v += inv_mass_b * impulse
+                des_b_a += rb.cross(impulse) * rb_b.inv_moment
+
+            # Friction
+
+            # Calculate friction coefficient
+            if a_none:
+                mu = rb_b.friction * rb_b.friction
+            elif b_none:
+                mu = rb_a.friction * rb_a.friction
+            else:
+                mu = min(rb_a.friction * rb_a.friction, rb_b.friction * rb_b.friction)
+
+            # Stop redundant friction calculations
+            if mu == 0:
+                continue
+
+            # Tangent vector
+            tangent = rv - rv.dot(col.normal) * col.normal
+            tangent.magnitude = 1
+
+            # Solve for magnitude to apply along the friction vector
+            jt = -rv.dot(tangent) * inv_sys_mass
+            jt /= len(col.contacts)
+
+            # Calculate friction impulse
+            if abs(jt) < j * mu:
+                friction_impulse = jt * tangent  # "Static friction"
+            else:
+                friction_impulse = -j * tangent * mu  # "Dynamic friction"
+
+            if not (a_none or rb_a.static):
+                des_a_v += inv_mass_a * friction_impulse
+                des_a_a += ra.cross(friction_impulse) * rb_a.inv_moment
+
+            if not (b_none or rb_b.static):
+                des_b_v += inv_mass_b * friction_impulse
+                des_b_a += rb.cross(friction_impulse) * rb_b.inv_moment
 
         # Position correction
-        correction = max(col.sep.magnitude - 0.01, 0) * inv_sys_mass * collision_norm
+        correction = max(col.penetration - 0.01, 0) * inv_sys_mass * col.normal
 
-        # Impulse Resolution
-
-        # Relative velocity
-        rv = (0 if rb_b_none else rb_b.velocity) - (0 if rb_a_none else rb_a.velocity)
-
-        if (vel_along_norm := rv.dot(collision_norm)) > 0:
-            return
-
-        # Calculate restitution
-        e = max(0 if rb_a_none else rb_a.bounciness, 0 if rb_b_none else rb_b.bounciness)
-
-        # Calculate impulse scalar
-        j = -(1 + e) * vel_along_norm * inv_sys_mass
-
-        # Apply the impulse
-        impulse = j * collision_norm
-
-        if not (rb_a_none or rb_a.static):
+        if not (a_none or rb_a.static):
             rb_a.gameobj.pos -= inv_mass_a * correction * rb_a.pos_correction
-            rb_a.velocity -= inv_mass_a * impulse
+            rb_a.velocity -= des_a_v
+            rb_a.ang_vel += des_a_a
 
-        if not (rb_b_none or rb_b.static):
+        if not (b_none or rb_b.static):
             rb_b.gameobj.pos += inv_mass_b * correction * rb_b.pos_correction
-            rb_b.velocity += inv_mass_b * impulse
-
-        # Friction
-
-        # Calculate friction coefficient
-        if rb_a_none:
-            mu = rb_b.friction * rb_b.friction
-        elif rb_b_none:
-            mu = rb_a.friction * rb_a.friction
-        else:
-            mu = min(rb_a.friction * rb_a.friction, rb_b.friction * rb_b.friction)
-
-        # Stop redundant friction calculations
-        if mu == 0:
-            return
-
-        # Tangent vector
-        tangent = rv - rv.dot(collision_norm) * collision_norm
-        tangent.magnitude = 1
-
-        # Solve for magnitude to apply along the friction vector
-        jt = -rv.dot(tangent) * inv_sys_mass
-
-        # Calculate friction impulse
-        if abs(jt) < j * mu:
-            friction_impulse = jt * tangent  # "Static friction"
-        else:
-            friction_impulse = -j * tangent * mu  # "Dynamic friction"
-
-        if not (rb_a_none or rb_a.static):
-            rb_a.velocity -= inv_mass_a * friction_impulse
-
-        if not (rb_b_none or rb_b.static):
-            rb_b.velocity += inv_mass_b * friction_impulse
+            rb_b.velocity += des_b_v
+            rb_b.ang_vel -= des_b_a